Support pg_* caps on pure cgroup2 systems
This patch has been prepared on a sles15sp6 cluster.

Don't set the subtree_control of the leaf cgroup.

When we are about to create a new leaf, set cgroup.subtree_control of
the parent, so the new child inherits a proper cgroup.controllers.

Don't add the CRM process pid in the non-leaf cgroup, to not block
setting of cgroup.subtree_control.

Note some changes:

* pg_vmem_limit can no longer be set alone as memory.swap.max is set to
  pg_vmem_limit-pg_mem_limit
* pg_mem_limit sets memory.max
* pg_cpu_quota can no longer set absolute values like "2000" or
  "2000@all". cgroup2 cpu.max accept a arbitrary period instead of the
  cfs timeslice duration, so we hardcoded a 200000 period and require
  percentile expressions like "10%" "10%all" or "10%2"
* pg_mem_oom_control and pg_mem_swapiness are ignored, as cgroup2 has no
  related tunables.