Change the task last run on-disk data cache to store more info
The <var>/<path>/<rid>/last_run_retcode file contained only the exitcode
and the last run date was inferred from the file modification time.

Now we want to store the last run session_id, for the user to correlate
with the action logs.

So use a json file to store all info. Maintain backward compat.

Example resource status with this patch:

        "task#2": {
            "info": {
                "at": "2025-06-26T18:06:41.732364+02:00",
                "exitcode": 7,
                "session_id": "fd84d838-f827-4eaf-ab11-a6ff1525634b"
            },
            "label": "task.host",
            "log": [
                "warn: last run exitcode: 7"
            ],
            "optional": true,
            "status": "n/a",
            "type": "task.host"
        }