Don't exit with code 0 when a *scoped* stop action has app issues
The agent avoids exiting with a non-zero code on app stop errors because
we want the stop action to continue to the fs stop, which can fix things
using `fuser -kmv <mnt>`. This policy maximizes chances of a svc switch
success, even when app stoppers are not reliable.

This patch just limit the app stop error bypass to the case where no
--rid --tag --subset commandline options were used with the stop action.