Add a configurable delay after a container.kvm is considered operational
The new qga_operational_delay keyword has a 10s default value.

When the container is operational, the start encap command will follow
very soon, and may fail because the operating system is still activating
subsystems.

For example we are already aware of sssd and dockerd dependencies in real
life situations. But we can not guess all dependencies of a particular
service and we don't want to maintain all deps testing involved.